DDSP_SVC_KOR_master/requirements.txt
protobuf==3.19.1
einops
omegaconf==2.0.5
fairseq==0.12.2
hydra-core==1.0.7
flask
flask_cors
gin
gin_config
librosa
local_attention
matplotlib
numpy
praat-parselmouth
pyworld
PyYAML
resampy
scikit_learn
scipy
SoundFile
tensorboard
torchcrepe
tqdm
transformers
wave
pysimplegui
sounddevice
gradio
jupyterlab